requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
